@@ -1,22 +1,23 @@
 {
-  "name": "developerfolio",
+  "name": "Kaushal_PORT_NEW",
   "version": "0.1.0",
   "private": true,
   "homepage": "https://kaushalyadv.github.io/Kaushal_PORT_NEW",
   "dependencies": {
+    "@babel/core": "^7.28.0",
+    "@react-three/drei": "^10.5.1",
+    "@react-three/fiber": "^9.2.0",
+    "babel-loader": "^10.0.0",
     "colorthief": "^2.3.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
+    "expo": "^53.0.20",
+    "flubber": "^0.4.2",
     "jest-canvas-mock": "^2.3.1",
     "lottie-react": "^2.4.0",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
+    "lottie-web": "^5.13.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-easy-emoji": "^1.8.1",
-    "react-headroom": "^3.0.0",
-    "react-reveal": "^1.2.2",
-    "react-scripts": "^0.0.0",
-    "react-test-renderer": "^16.10.2",
-    "react-twitter-embed": "3.0.3",
+    "react-scripts": "^5.0.1",
     "sass": "^1.32.8"
   },
   "scripts": {
