@@ -1,5 +1,4 @@
 import React, {useContext} from "react";
-import {Fade} from "react-reveal";
 import emoji from "react-easy-emoji";
 import "./Greeting.scss";
 import landingPerson from "../../assets/lottie/landingPerson";
@@ -15,56 +14,53 @@ export default function Greeting() {
     return null;
   }
   return (
-    <Fade bottom duration={1000} distance="40px">
-      <div className="greet-main" id="greeting">
-        <div className="greeting-main">
-          <div className="greeting-text-div">
-            <div>
-              <h1
-                className={isDark ? "dark-mode greeting-text" : "greeting-text"}
-              >
-                {" "}
-                {greeting.title}{" "}
-                <span className="wave-emoji">{emoji("👋")}</span>
-              </h1>
-              <p
-                className={
-                  isDark
-                    ? "dark-mode greeting-text-p"
-                    : "greeting-text-p subTitle"
-                }
-              >
-                {greeting.subTitle}
-              </p>
-              <div id="resume" className="empty-div"></div>
-              <SocialMedia />
-              <div className="button-greeting-div">
-                <Button text="Contact me" href="#contact" />
-                {greeting.resumeLink && (
-                  <a
-                    href={require("./Final_resume.pdf")}
-                    download="Resume.pdf"
-                    className="download-link-button"
-                  >
-                    <Button text="Download my resume" link="https://docs.google.com/presentation/d/14uF7NmhsXxVjML5RhtkP6UHCqZto3VUo/edit?usp=sharing&ouid=115960713439289799000&rtpof=true&sd=true" />
-
-                  </a>
-                )}
-              </div>
+    <div className="greet-main" id="greeting">
+      <div className="greeting-main">
+        <div className="greeting-text-div">
+          <div>
+            <h1
+              className={isDark ? "light-mode greeting-text" : "greeting-text"}
+            >
+              {" "}
+              {greeting.title}{" "}
+              <span className="wave-emoji">{emoji("👋")}</span>
+            </h1>
+            <p
+              className={
+                isDark
+                  ? "light-mode greeting-text-p"
+                  : "greeting-text-p subTitle"
+              }
+            >
+              {greeting.subTitle}
+            </p>
+            <div id="resume" className="empty-div"></div>
+            <SocialMedia />
+            <div className="button-greeting-div">
+              <Button text="Contact me" href="#contact" />
+              {greeting.resumeLink && (
+                <a
+                  href={require("./Final_resume.pdf")}
+                  download="Resume.pdf"
+                  className="download-link-button"
+                >
+                  <Button text="Download my resume" link="https://docs.google.com/presentation/d/14uF7NmhsXxVjML5RhtkP6UHCqZto3VUo/edit?usp=sharing&ouid=115960713439289799000&rtpof=true&sd=true" />
+                </a>
+              )}
             </div>
           </div>
-          <div className="greeting-image-div">
-            {illustration.animated ? (
-              <DisplayLottie animationData={landingPerson} />
-            ) : (
-              <img
-                alt="man sitting on table"
-                src={require("../../assets/images/manOnTable.svg")}
-              ></img>
-            )}
-          </div>
+        </div>
+        <div className="greeting-image-div">
+          {illustration.animated ? (
+            <DisplayLottie animationData={landingPerson} />
+          ) : (
+            <img
+              alt="man sitting on table"
+              src={require("../../assets/images/manOnTable.svg")}
+            ></img>
+          )}
         </div>
       </div>
-    </Fade>
+    </div>
   );
 }
