@@ -2,7 +2,6 @@ import React, {useContext} from "react";
 import "./Skills.scss";
 import SoftwareSkill from "../../components/softwareSkills/SoftwareSkill";
 import {illustration, skillsSection} from "../../portfolio";
-import {Fade} from "react-reveal";
 import codingPerson from "../../assets/lottie/codingPerson";
 import DisplayLottie from "../../components/displayLottie/DisplayLottie";
 import StyleContext from "../../contexts/StyleContext";
@@ -13,9 +12,8 @@ export default function Skills() {
     return null;
   }
   return (
-    <div className={isDark ? "dark-mode main" : "main"} id="skills">
+    <div className={isDark ? "light-mode main" : "main"} id="skills">
       <div className="skills-main-div">
-        <Fade left duration={1000}>
           <div className="skills-image-div">
             {illustration.animated ? (
               <DisplayLottie animationData={codingPerson} />
@@ -26,18 +24,16 @@ export default function Skills() {
               ></img>
             )}
           </div>
-        </Fade>
-        <Fade right duration={1000}>
           <div className="skills-text-div">
             <h1
-              className={isDark ? "dark-mode skills-heading" : "skills-heading"}
+              className={isDark ? "light-mode skills-heading" : "skills-heading"}
             >
               {skillsSection.title}{" "}
             </h1>
             <p
               className={
                 isDark
-                  ? "dark-mode subTitle skills-text-subtitle"
+                  ? "light-mode subTitle skills-text-subtitle"
                   : "subTitle skills-text-subtitle"
               }
             >
@@ -51,7 +47,7 @@ export default function Skills() {
                     key={i}
                     className={
                       isDark
-                        ? "dark-mode subTitle skills-text"
+                        ? "light-mode subTitle skills-text"
                         : "subTitle skills-text"
                     }
                   >
@@ -61,7 +57,6 @@ export default function Skills() {
               })}
             </div>
           </div>
-        </Fade>
       </div>
     </div>
   );
