@@ -7,7 +7,7 @@ import StyleContext from "../../contexts/StyleContext";
 export default function SplashScreen() {
   const {isDark} = useContext(StyleContext);
   return (
-    <div className={isDark ? "dark-mode splash-container" : "splash-container"}>
+    <div className={isDark ? "light-mode splash-container" : "splash-container"}>
       <div className="splash-animation-container">
         <DisplayLottie animationData={splashScreen.animation} />
       </div>
